# Remove the Active → Faulted transition