--- a/Cargo.toml
+++ b/Cargo.toml
@@ -53,6 +53,7 @@ hex = "0.4"
 http = "0.2.12"
 httptest = "0.15.5"
 human_bytes = "0.4.3"
+humantime = "2.1.0"
 hyper = { version = "0.14", features = [ "full" ] }
 hyper-staticfile = "0.9"
 indicatif = { version = "0.17.8", features = ["rayon"] }