This PR removes the Active → Faulted transition when too many IO operations are in flight.

See #1252 for details on why this is desirable!

There are a bunch of changes that come along for the ride:

- The backpressure equation is changed from quadratic to the form `1 / (1 - f)`, i.e. it now goes to infinity at our desired maximum value.  These maximum values are set at 2x the fault for the Offline → Faulted transition.
- Computing backpressure is now a standalone function on `BackpressureConfig`, so it can be unit tested
- Unit tests are added to confirm backpressure properties
  - The `Offline` → `Faulted` transition happens at a reasonable timescale (> 1 sec, < 2 minutes)
  - Backpressure goes to ∞ as we approach these limits
- Backpressure settings are tweaked based on these new requirements